Hardware main to treat all instructions as LWI, running cmptest.
cmptest is:
0i0
ii1
i1i

So this is expected:

A0 0 0 0
A1 i i 1
A2 0 1 i
PC 0 1 i (and repeating)

And that's what I see. But in graphing A1, from
63-70 µs the voltage goes to 0, until it is reset
back to -5 V by the next LWI instruction. In
cmptest without LWI hardwired, V(A1) goes 0
at this time but doesn't return because the next
instruction is not LWI. What is making this voltage
go up for 7 µs?