Fix swrom board, by swapping pins 3 and 4 on the footprint.
Generate CAM files swrom-fixed, submitted in order http://jeff.tk/wiki/Trinary/Parts#Purchase_13 .

Note that the modified footprint is not included here, because it will be
incorrect if a circuit is now generated with pcb.py, since pins 3 and 4
were swapped in sp3t-?.asy already, so the problem would persist. Use the
same footprint in new circuits.